+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+* 2x.x.x
+  - Fixed PowerMethod for square/non-square, complex/float matrices with stopping criterion.
+  - CofR image_sharpness improved for large datasets
+  - Geometry alignmentment fix for 2D datasets
+  - CGLS update for sapyb to enable complex data
+  - added sapyb and deprecated axpby
+
+* 21.3.1
+  - Added matplotlib version dependency to conda recipe
+  - Fixed TIGRE wrappers for geometry with a virtual detector
+  - Fixed TIGRE wrappers for cone-beam geometry with tilted rotation axis
+
 * 21.3.0
   - Accelerated PDHG which handles strong convexity of functions
   - TotalVariation Function handles SIRF ImageData